+++ wildlifelicensing/apps/customer_management/templates/wl/officer_edit_customer_profile.html
@@ -10,8 +10,8 @@
             <div class="col-md-12">
                 <ol class="breadcrumb wl-breadcrumbs">
                     <li><a href="{% url 'home' %}">Home</a></li>
-                    <li><a href="{% url 'customer_management:lookup_customer' %}">Customer Lookup</a></li>
-                    <li><a href="{% url 'customer_management:view_customer' customer.pk %}">View Customer</a></li>
+                    <li><a href="{% url 'wl_customer_management:lookup_customer' %}">Customer Lookup</a></li>
+                    <li><a href="{% url 'wl_customer_management:view_customer' customer.pk %}">View Customer</a></li>
                     <li class="active">{% if profile_form.user.value %}Edit{% else %}Add{% endif %} Profile</li>
                 </ol>
             </div>
